Allow test fixtures to be loaded in postgresql
In MySQL, boolean columns are one-bit integers that store the
value 0 for false and 1 for true.  In PostgreSQL, they are
boolean columns that store true and false.

The cards.yml in db/bootsrap has the value 'false' for the
trash column in all the card records, which works fine in both
databases.  The cards.yml in test/fixtures has the value '0' for
the trash column, which only works in MySQL.

Use false instead!