list and listed_by cardtypes
A way to establish has_many relations. The pattern is as follows:
Let A,B be cards of type letter and 1,2 cards of type number.
If you create list cards
 - A+numbers -> [[1]]\n[[2]]
 - B+numbers -> [[1]]
then the following listed_by cards will contain the related letters:
 - 1+letters -> [[A]]\n[[B]]
 - 2+letters -> [[A]]]

The list cards have to be a real cards. The listed_by cards are virtual. Their content
is cached and gets updated after every update of an related list card.